requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1


# third party packages
tensorflow==2.13.0
pandas
numpy
scipy
matplotlib
mlflow
Flask
tqdm
notebook
gdown
python-box
ensure